fix：修复了数据库表的问题
